+++ .env.example
@@ -1,8 +1,9 @@
 NODE_ENV=development
-MONGOHQ_URL=
-MONGOHQ_URL_TEST=
+MONGOHQ_URL=mongodb://localhost:27017/cfh
+MONGOHQ_URL_TEST=mongodb://localhost:27017/cfh_test
+MONGOHQ_URL_STAGING=mongodb://staging-address/gondor-cfh-staging
+MONGOHQ_URL_PROD=mongodb://production-address/gondor-cfh
 SECRET=
-
 APP_URL=http://localhost:3000
 TWITTER_CONSUMER_KEY={key}
 TWITTER_CONSUMER_SECRET={secret}
