+++ package.json
@@ -23,7 +23,14 @@
     "merge:lcov": "cat ./coverage/*/lcov.info > ./coverage/lcov.info",
     "coverage": "npm run merge:lcov && cat ./coverage/lcov.info | coveralls",
     "heroku-postbuild": "npm run build",
-    "watch:scss": "gulp watchSCSS"
+    "migrate": "cross-env NODE_ENV=development ./node_modules/.bin/migrate-mongo up",
+    "migrate:test": "cross-env NODE_ENV=test ./node_modules/.bin/migrate-mongo up",
+    "migrate:staging": "cross-env NODE_ENV=staging ./node_modules/.bin/migrate-mongo up",
+    "migrate:prod": "cross-env NODE_ENV=production ./node_modules/.bin/migrate-mongo up",
+    "migrate:down": "cross-env NODE_ENV=development ./node_modules/.bin/migrate-mongo down",
+    "migrate:test-down": "cross-env NODE_ENV=test ./node_modules/.bin/migrate-mongo down",
+    "migrate:staging-down": "cross-env NODE_ENV=staging ./node_modules/.bin/migrate-mongo down",
+    "migrate:prod-down": "cross-env NODE_ENV=production ./node_modules/.bin/migrate-mongo down"
   },
   "dependencies": {
     "async": "~0.2.9",
@@ -42,6 +49,7 @@
     "jsonwebtoken": "^8.3.0",
     "karma-babel-preprocessor": "^7.0.0",
     "mean-logger": "^0.0.1",
+    "migrate-mongo": "^3.0.2",
     "mongoose": "4.13.14",
     "mongoose-unique-validator": "^2.0.1",
     "nodemailer": "^4.6.7",
