+++ seed_data/answers-data.js
@@ -1,0 +1,3107 @@
+/* eslint-disable */
+const answers = [
+  {
+    expansion: 'Base',
+    id: 2,
+    official: true, region: 'Africa',
+    text: "Michelle Obama's arms."
+  },
+  {
+    expansion: 'Base',
+    id: 5,
+    official: true, region: 'Africa',
+    text: 'Getting so angry that you pop a boner.'
+  },
+  {
+    expansion: 'Base',
+    id: 6,
+    official: true, region: 'Africa',
+    text: 'Tasteful sideboob.'
+  },
+  {
+    expansion: 'Base',
+    id: 8,
+    official: true, region: 'Africa',
+    text: 'Two midgets shitting into a bucket.'
+  },
+  {
+    expansion: 'Base',
+    id: 9,
+    official: true, region: 'Africa',
+    text: 'MechaHitler.'
+  },
+  {
+    expansion: 'Base',
+    id: 11,
+    official: true, region: 'Africa',
+    text: 'A disappointing birthday party.'
+  },
+  {
+    expansion: 'Base',
+    id: 12,
+    official: true, region: 'Africa',
+    text: 'Puppies!'
+  },
+  {
+    expansion: 'Base',
+    id: 14,
+    official: true, region: 'Africa',
+    text: "Guys who don't call."
+  },
+  {
+    expansion: 'Base',
+    id: 15,
+    official: true, region: 'Africa',
+    text: 'Racially-biased SAT questions.'
+  },
+  {
+    expansion: 'Base',
+    id: 18,
+    official: true, region: 'Africa',
+    text: 'Being on fire.'
+  },
+  {
+    expansion: 'Base',
+    id: 19,
+    official: true, region: 'Africa',
+    text: 'A lifetime of sadness.'
+  },
+  {
+    expansion: 'Base',
+    id: 23,
+    official: true, region: 'Africa',
+    text: 'Glenn Beck catching his scrotum on a curtain hook.'
+  },
+  {
+    expansion: 'Base',
+    id: 24,
+    official: true, region: 'Africa',
+    text: 'The Rapture.'
+  },
+  {
+    expansion: 'Base',
+    id: 26,
+    official: true, region: 'Africa',
+    text: 'Crippling debt.'
+  },
+  {
+    expansion: 'Base',
+    id: 27,
+    official: true, region: 'Africa',
+    text: 'Eugenics.'
+  },
+  {
+    expansion: 'Base',
+    id: 28,
+    official: true, region: 'Africa',
+    text: 'Exchanging pleasantries.'
+  },
+  {
+    expansion: 'Base',
+    id: 29,
+    official: true, region: 'Africa',
+    text: 'Dying of dysentery.'
+  },
+  {
+    expansion: 'Base',
+    id: 30,
+    official: true, region: 'Africa',
+    text: 'Roofies.'
+  },
+  {
+    expansion: 'Base',
+    id: 32,
+    official: true, region: 'Africa',
+    text: 'The forbidden fruit.'
+  },
+  {
+    expansion: 'Base',
+    id: 33,
+    official: true, region: 'Africa',
+    text: 'Republicans.'
+  },
+  {
+    expansion: 'Base',
+    id: 34,
+    official: true, region: 'Africa',
+    text: 'The Big Bang.'
+  },
+  {
+    expansion: 'Base',
+    id: 36,
+    official: true, region: 'Africa',
+    text: 'Amputees.'
+  },
+  {
+    expansion: 'Base',
+    id: 37,
+    official: true, region: 'Africa',
+    text: 'Men.'
+  },
+  {
+    expansion: 'Base',
+    id: 39,
+    official: true, region: 'Africa',
+    text: 'Kim Jong-il.'
+  },
+  {
+    expansion: 'Base',
+    id: 40,
+    official: true, region: 'Africa',
+    text: 'Concealing a boner'
+  },
+  {
+    expansion: 'Base',
+    id: 41,
+    official: true, region: 'Africa',
+    text: 'Agriculture.'
+  },
+  {
+    expansion: 'Base',
+    id: 42,
+    official: true, region: 'Africa',
+    text: 'Glenn Beck being harried by a swarm of buzzards.'
+  },
+  {
+    expansion: 'Base',
+    id: 43,
+    official: true, region: 'Africa',
+    text: 'Making a pouty face.'
+  },
+  {
+    expansion: 'Base',
+    id: 44,
+    official: true, region: 'Africa',
+    text: 'A salty surprise.'
+  },
+  {
+    expansion: 'Base',
+    id: 45,
+    official: true, region: 'Africa',
+    text: 'The Jews.'
+  },
+  {
+    expansion: 'Base',
+    id: 46,
+    official: true, region: 'Africa',
+    text: 'Charisma.'
+  },
+  {
+    expansion: 'Base',
+    id: 47,
+    official: true, region: 'Africa',
+    text: 'YOU MUST CONSTRUCT ADDITIONAL PYLONS.'
+  },
+  {
+    expansion: 'Base',
+    id: 50,
+    official: true, region: 'Africa',
+    text: 'A drive-by shooting.'
+  },
+  {
+    expansion: 'Base',
+    id: 51,
+    official: true, region: 'Africa',
+    text: 'Ronald Reagan.'
+  },
+  {
+    expansion: 'Base',
+    id: 52,
+    official: true, region: 'Africa',
+    text: "Morgan Freeman's voice."
+  },
+  {
+    expansion: 'Base',
+    id: 53,
+    official: true, region: 'Africa',
+    text: 'Breaking out into song and dance.'
+  },
+  {
+    expansion: 'Base',
+    id: 54,
+    official: true, region: 'Africa',
+    text: 'Jewish fraternities.'
+  },
+  {
+    expansion: 'Base',
+    id: 57,
+    official: true, region: 'Africa',
+    text: 'Hormone injections.'
+  },
+  {
+    expansion: 'Base',
+    id: 58,
+    official: true, region: 'Africa',
+    text: 'All-you-can-eat shrimp for $4.99.'
+  },
+  {
+    expansion: 'Base',
+    id: 59,
+    official: true, region: 'Africa',
+    text: 'Incest.'
+  },
+  {
+    expansion: 'Base',
+    id: 60,
+    official: true, region: 'Africa',
+    text: 'Scalping.'
+  },
+  {
+    expansion: 'Base',
+    id: 61,
+    official: true, region: 'Africa',
+    text: 'Soup that is too hot.'
+  },
+  {
+    expansion: 'Base',
+    id: 62,
+    official: true, region: 'Africa',
+    text: 'The Übermensch.'
+  },
+  {
+    expansion: 'Base',
+    id: 63,
+    official: true, region: 'Africa',
+    text: 'Nazis.'
+  },
+  {
+    expansion: 'Base',
+    id: 64,
+    official: true, region: 'Africa',
+    text: 'Tom Cruise.'
+  },
+  {
+    expansion: 'Base',
+    id: 65,
+    official: true, region: 'Africa',
+    text: 'Stifling a giggle at the mention of Hutus and Tutsis.'
+  },
+  {
+    expansion: 'Base',
+    id: 67,
+    official: true, region: 'Africa',
+    text: 'The Hustle.'
+  },
+  {
+    expansion: 'Base',
+    id: 68,
+    official: true, region: 'Africa',
+    text: 'A Super Soaker™ full of cat pee.'
+  },
+  {
+    expansion: 'Base',
+    id: 69,
+    official: true, region: 'Africa',
+    text: 'Figgy pudding.'
+  },
+  {
+    expansion: 'Base',
+    id: 70,
+    official: true, region: 'Africa',
+    text: 'Object permanence.'
+  },
+  {
+    expansion: 'Base',
+    id: 71,
+    official: true, region: 'Africa',
+    text: 'Consultants.'
+  },
+  {
+    expansion: 'Base',
+    id: 72,
+    official: true, region: 'Africa',
+    text: 'Intelligent design.'
+  },
+  {
+    expansion: 'Base',
+    id: 73,
+    official: true, region: 'Africa',
+    text: 'Nocturnal emissions.'
+  },
+  {
+    expansion: 'Base',
+    id: 74,
+    official: true, region: 'Africa',
+    text: 'Uppercuts.'
+  },
+  {
+    expansion: 'Base',
+    id: 75,
+    official: true, region: 'Africa',
+    text: 'Being marginalized.'
+  },
+  {
+    expansion: 'Base',
+    id: 76,
+    official: true, region: 'Africa',
+    text: 'The profoundly handicapped.'
+  },
+  {
+    expansion: 'Base',
+    id: 77,
+    official: true, region: 'Africa',
+    text: 'Obesity.'
+  },
+  {
+    expansion: 'Base',
+    id: 78,
+    official: true, region: 'Africa',
+    text: 'Chutzpah.'
+  },
+  {
+    expansion: 'Base',
+    id: 79,
+    official: true, region: 'Africa',
+    text: 'Unfathomable stupidity.'
+  },
+  {
+    expansion: 'Base',
+    id: 80,
+    official: true, region: 'Africa',
+    text: 'Repression.'
+  },
+  {
+    expansion: 'Base',
+    id: 81,
+    official: true, region: 'Africa',
+    text: 'Attitude.'
+  },
+  {
+    expansion: 'Base',
+    id: 82,
+    official: true, region: 'Africa',
+    text: 'Passable transvestites.'
+  },
+  {
+    expansion: 'Base',
+    id: 84,
+    official: true, region: 'Africa',
+    text: 'The American Dream'
+  },
+  {
+    expansion: 'Base',
+    id: 85,
+    official: true, region: 'Africa',
+    text: 'Child beauty pageants.'
+  },
+  {
+    expansion: 'Base',
+    id: 86,
+    official: true, region: 'Africa',
+    text: 'Puberty.'
+  },
+  {
+    expansion: 'Base',
+    id: 88,
+    official: true, region: 'Africa',
+    text: 'The folly of man.'
+  },
+  {
+    expansion: 'Base',
+    id: 89,
+    official: true, region: 'Africa',
+    text: 'Nickelback.'
+  },
+  {
+    expansion: 'Base',
+    id: 90,
+    official: true, region: 'Africa',
+    text: 'Swooping.'
+  },
+  {
+    expansion: 'Base',
+    id: 91,
+    official: true, region: 'Africa',
+    text: 'Goats eating cans.'
+  },
+  {
+    expansion: 'Base',
+    id: 92,
+    official: true, region: 'Africa',
+    text: 'The KKK.'
+  },
+  {
+    expansion: 'Base',
+    id: 93,
+    official: true, region: 'Africa',
+    text: 'Kamikaze pilots.'
+  },
+  {
+    expansion: 'Base',
+    id: 94,
+    official: true, region: 'Africa',
+    text: 'Horrifying laser hair removal accidents.'
+  },
+  {
+    expansion: 'Base',
+    id: 95,
+    official: true, region: 'Africa',
+    text: 'Adderall™.'
+  },
+  {
+    expansion: 'Base',
+    id: 96,
+    official: true, region: 'Africa',
+    text: 'A look-see.'
+  },
+  {
+    expansion: 'Base',
+    id: 97,
+    official: true, region: 'Africa',
+    text: 'Doing the right thing.'
+  },
+  {
+    expansion: 'Base',
+    id: 98,
+    official: true, region: 'Africa',
+    text: 'The taint; the grundle; the fleshy fun-bridge.'
+  },
+  {
+    expansion: 'Base',
+    id: 99,
+    official: true, region: 'Africa',
+    text: 'Lactation.'
+  },
+  {
+    expansion: 'Base',
+    id: 100,
+    official: true, region: 'Africa',
+    text: 'Pabst Blue Ribbon.'
+  },
+  {
+    expansion: 'Base',
+    id: 101,
+    official: true, region: 'Africa',
+    text: 'Powerful thighs.'
+  },
+  {
+    expansion: 'Base',
+    id: 102,
+    official: true, region: 'Africa',
+    text: 'Saxophone solos.'
+  },
+  {
+    expansion: 'Base',
+    id: 104,
+    official: true, region: 'Africa',
+    text: 'A middle-aged man on roller skates.'
+  },
+  {
+    expansion: 'Base',
+    id: 105,
+    official: true, region: 'Africa',
+    text: 'A foul mouth.'
+  },
+  {
+    expansion: 'Base',
+    id: 106,
+    official: true, region: 'Africa',
+    text: 'The thing that electrocutes your abs.'
+  },
+  {
+    expansion: 'Base',
+    id: 107,
+    official: true, region: 'Africa',
+    text: 'Heteronormativity.'
+  },
+  {
+    expansion: 'Base',
+    id: 108,
+    official: true, region: 'Africa',
+    text: 'Cuddling.'
+  },
+  {
+    expansion: 'Base',
+    id: 109,
+    official: true, region: 'Africa',
+    text: 'Coat hanger abortions.'
+  },
+  {
+    expansion: 'Base',
+    id: 110,
+    official: true, region: 'Africa',
+    text: 'A big hoopla about nothing.'
+  },
+  {
+    expansion: 'Base',
+    id: 111,
+    official: true, region: 'Africa',
+    text: 'Boogers.'
+  },
+  {
+    expansion: 'Base',
+    id: 112,
+    official: true, region: 'Africa',
+    text: 'A hot mess.'
+  },
+  {
+    expansion: 'Base',
+    id: 113,
+    official: true, region: 'Africa',
+    text: 'Raptor attacks.'
+  },
+  {
+    expansion: 'Base',
+    id: 115,
+    official: true, region: 'Africa',
+    text: 'Fear itself.'
+  },
+  {
+    expansion: 'Base',
+    id: 116,
+    official: true, region: 'Africa',
+    text: 'Bees?'
+  },
+  {
+    expansion: 'Base',
+    id: 117,
+    official: true, region: 'Africa',
+    text: 'Getting drunk on mouthwash.'
+  },
+  {
+    expansion: 'Base',
+    id: 118,
+    official: true, region: 'Africa',
+    text: 'Sniffing glue.'
+  },
+  {
+    expansion: 'Base',
+    id: 119,
+    official: true, region: 'Africa',
+    text: 'Oversized lollipops.'
+  },
+  {
+    expansion: 'Base',
+    id: 120,
+    official: true, region: 'Africa',
+    text: 'An icepick lobotomy.'
+  },
+  {
+    expansion: 'Base',
+    id: 121,
+    official: true, region: 'Africa',
+    text: 'Being rich.'
+  },
+  {
+    expansion: 'Base',
+    id: 123,
+    official: true, region: 'Africa',
+    text: 'Teaching a robot to love.'
+  },
+  {
+    expansion: 'Base',
+    id: 124,
+    official: true, region: 'Africa',
+    text: "Women's suffrage."
+  },
+  {
+    expansion: 'Base',
+    id: 125,
+    official: true, region: 'Africa',
+    text: 'Me time.'
+  },
+  {
+    expansion: 'Base',
+    id: 126,
+    official: true, region: 'Africa',
+    text: 'The heart of a child.'
+  },
+  {
+    expansion: 'Base',
+    id: 127,
+    official: true, region: 'Africa',
+    text: 'Smallpox blankets.'
+  },
+  {
+    expansion: 'Base',
+    id: 129,
+    official: true, region: 'Africa',
+    text: 'John Wilkes Booth.'
+  },
+  {
+    expansion: 'Base',
+    id: 130,
+    official: true, region: 'Africa',
+    text: 'The glass ceiling.'
+  },
+  {
+    expansion: 'Base',
+    id: 131,
+    official: true, region: 'Africa',
+    text: 'Sarah Palin.'
+  },
+  {
+    expansion: 'Base',
+    id: 133,
+    official: true, region: 'Africa',
+    text: 'Yeast.'
+  },
+  {
+    expansion: 'Base',
+    id: 135,
+    official: true, region: 'Africa',
+    text: 'Parting the Red Sea.'
+  },
+  {
+    expansion: 'Base',
+    id: 136,
+    official: true, region: 'Africa',
+    text: 'A Bop It™.'
+  },
+  {
+    expansion: 'Base',
+    id: 137,
+    official: true, region: 'Africa',
+    text: 'Michael Jackson.'
+  },
+  {
+    expansion: 'Base',
+    id: 138,
+    official: true, region: 'Africa',
+    text: 'Team-building exercises.'
+  },
+  {
+    expansion: 'Base',
+    id: 139,
+    official: true, region: 'Africa',
+    text: 'Harry Potter erotica.'
+  },
+  {
+    expansion: 'Base',
+    id: 140,
+    official: true, region: 'Africa',
+    text: 'Authentic Mexican cuisine.'
+  },
+  {
+    expansion: 'Base',
+    id: 143,
+    official: true, region: 'Africa',
+    text: 'Grandma.'
+  },
+  {
+    expansion: 'Base',
+    id: 144,
+    official: true, region: 'Africa',
+    text: 'Not giving a shit about the Third World.'
+  },
+  {
+    expansion: 'Base',
+    id: 145,
+    official: true, region: 'Africa',
+    text: 'Licking things to claim them as your own.'
+  },
+  {
+    expansion: 'Base',
+    id: 146,
+    official: true, region: 'Africa',
+    text: 'Genghis Khan.'
+  },
+  {
+    expansion: 'Base',
+    id: 147,
+    official: true, region: 'Africa',
+    text: 'The hardworking Mexican.'
+  },
+  {
+    expansion: 'Base',
+    id: 148,
+    official: true, region: 'Africa',
+    text: 'RoboCop.'
+  },
+  {
+    expansion: 'Base',
+    id: 149,
+    official: true, region: 'Africa',
+    text: 'My relationship status.'
+  },
+  {
+    expansion: 'Base',
+    id: 150,
+    official: true, region: 'Africa',
+    text: 'Scrubbing under the folds.'
+  },
+  {
+    expansion: 'Base',
+    id: 152,
+    official: true, region: 'Africa',
+    text: 'Horse meat.'
+  },
+  {
+    expansion: 'Base',
+    id: 153,
+    official: true, region: 'Africa',
+    text: 'Sunshine and rainbows.'
+  },
+  {
+    expansion: 'Base',
+    id: 155,
+    official: true, region: 'Africa',
+    text: 'Barack Obama.'
+  },
+  {
+    expansion: 'Base',
+    id: 156,
+    official: true, region: 'Africa',
+    text: 'Spontaneous human combustion.'
+  },
+  {
+    expansion: 'Base',
+    id: 157,
+    official: true, region: 'Africa',
+    text: 'Natural selection.'
+  },
+  {
+    expansion: 'Base',
+    id: 158,
+    official: true, region: 'Africa',
+    text: 'Mouth herpes.'
+  },
+  {
+    expansion: 'Base',
+    id: 159,
+    official: true, region: 'Africa',
+    text: 'Flash flooding.'
+  },
+  {
+    expansion: 'Base',
+    id: 160,
+    official: true, region: 'Africa',
+    text: 'Goblins.'
+  },
+  {
+    expansion: 'Base',
+    id: 161,
+    official: true, region: 'Africa',
+    text: 'A monkey smoking a cigar.'
+  },
+  {
+    expansion: 'Base',
+    id: 162,
+    official: true, region: 'Africa',
+    text: 'Spectacular abs.'
+  },
+  {
+    expansion: 'Base',
+    id: 163,
+    official: true, region: 'Africa',
+    text: 'A good sniff.'
+  },
+  {
+    expansion: 'Base',
+    id: 165,
+    official: true, region: 'Africa',
+    text: 'The Three-Fifths compromise.'
+  },
+  {
+    expansion: 'Base',
+    id: 166,
+    official: true, region: 'Africa',
+    text: 'Pedophiles.'
+  },
+  {
+    expansion: 'Base',
+    id: 168,
+    official: true, region: 'Africa',
+    text: 'Being fabulous.'
+  },
+  {
+    expansion: 'Base',
+    id: 169,
+    official: true, region: 'Africa',
+    text: 'Mathletes.'
+  },
+  {
+    expansion: 'Base',
+    id: 170,
+    official: true, region: 'Africa',
+    text: 'Wearing underwear inside-out to avoid doing laundry.'
+  },
+  {
+    expansion: 'Base',
+    id: 171,
+    official: true, region: 'Africa',
+    text: 'Nipple blades.'
+  },
+  {
+    expansion: 'Base',
+    id: 172,
+    official: true, region: 'Africa',
+    text: 'An M. Night Shyamalan plot twist.'
+  },
+  {
+    expansion: 'Base',
+    id: 173,
+    official: true, region: 'Africa',
+    text: 'A bag of magic beans.'
+  },
+  {
+    expansion: 'Base',
+    id: 174,
+    official: true, region: 'Africa',
+    text: 'Vigorous jazz hands.'
+  },
+  {
+    expansion: 'Base',
+    id: 176,
+    official: true, region: 'Africa',
+    text: 'Skeletor.'
+  },
+  {
+    expansion: 'Base',
+    id: 177,
+    official: true, region: 'Africa',
+    text: 'Vikings.'
+  },
+  {
+    expansion: 'Base',
+    id: 178,
+    official: true, region: 'Africa',
+    text: 'Leaving an awkward voicemail.'
+  },
+  {
+    expansion: 'Base',
+    id: 179,
+    official: true, region: 'Africa',
+    text: 'Teenage pregnancy.'
+  },
+  {
+    expansion: 'Base',
+    id: 180,
+    official: true, region: 'Africa',
+    text: 'Dead parents.'
+  },
+  {
+    expansion: 'Base',
+    id: 181,
+    official: true, region: 'Africa',
+    text: 'Hot cheese.'
+  },
+  {
+    expansion: 'Base',
+    id: 183,
+    official: true, region: 'Africa',
+    text: 'A mopey zoo lion.'
+  },
+  {
+    expansion: 'Base',
+    id: 184,
+    official: true, region: 'Africa',
+    text: 'Assless chaps.'
+  },
+  {
+    expansion: 'Base',
+    id: 185,
+    official: true, region: 'Africa',
+    text: 'Riding off into the sunset.'
+  },
+  {
+    expansion: 'Base',
+    id: 186,
+    official: true, region: 'Africa',
+    text: "Lance Armstrong's missing testicle."
+  },
+  {
+    expansion: 'Base',
+    id: 187,
+    official: true, region: 'Africa',
+    text: 'Sweet, sweet vengeance.'
+  },
+  {
+    expansion: 'Base',
+    id: 189,
+    official: true, region: 'Africa',
+    text: 'Keg stands.'
+  },
+  {
+    expansion: 'Base',
+    id: 190,
+    official: true, region: 'Africa',
+    text: 'Darth Vader.'
+  },
+  {
+    expansion: 'Base',
+    id: 192,
+    official: true, region: 'Africa',
+    text: 'Necrophilia.'
+  },
+  {
+    expansion: 'Base',
+    id: 193,
+    official: true, region: 'Africa',
+    text: 'A really cool hat.'
+  },
+  {
+    expansion: 'Base',
+    id: 195,
+    official: true, region: 'Africa',
+    text: 'An Oedipus complex.'
+  },
+  {
+    expansion: 'Base',
+    id: 196,
+    official: true, region: 'Africa',
+    text: 'The Tempur-Pedic® Swedish Sleep System™.'
+  },
+  {
+    expansion: 'Base',
+    id: 197,
+    official: true, region: 'Africa',
+    text: 'Preteens.'
+  },
+  {
+    expansion: 'Base',
+    id: 198,
+    official: true, region: 'Africa',
+    text: 'Dick fingers.'
+  },
+  {
+    expansion: 'Base',
+    id: 199,
+    official: true, region: 'Africa',
+    text: 'A cooler full of organs.'
+  },
+  {
+    expansion: 'Base',
+    id: 200,
+    official: true, region: 'Africa',
+    text: 'Shapeshifters.'
+  },
+  {
+    expansion: 'Base',
+    id: 201,
+    official: true, region: 'Europe',
+    text: 'The Care Bear Stare.'
+  },
+  {
+    expansion: 'Base',
+    id: 202,
+    official: true, region: 'Europe',
+    text: 'Erectile dysfunction.'
+  },
+  {
+    expansion: 'Base',
+    id: 203,
+    official: true, region: 'Europe',
+    text: 'Keanu Reeves.'
+  },
+  {
+    expansion: 'Base',
+    id: 204,
+    official: true, region: 'Europe',
+    text: 'The Virginia Tech Massacre.'
+  },
+  {
+    expansion: 'Base',
+    id: 205,
+    official: true, region: 'Europe',
+    text: 'The Underground Railroad.'
+  },
+  {
+    expansion: 'Base',
+    id: 206,
+    official: true, region: 'Europe',
+    text: 'The chronic.'
+  },
+  {
+    expansion: 'Base',
+    id: 207,
+    official: true, region: 'Europe',
+    text: 'Queefing.'
+  },
+  {
+    expansion: 'Base',
+    id: 208,
+    official: true, region: 'Europe',
+    text: 'Heartwarming orphans.'
+  },
+  {
+    expansion: 'Base',
+    id: 209,
+    official: true, region: 'Europe',
+    text: 'A thermonuclear detonation.'
+  },
+  {
+    expansion: 'Base',
+    id: 210,
+    official: true, region: 'Europe',
+    text: 'Cheating in the Special Olympics.'
+  },
+  {
+    expansion: 'Base',
+    id: 211,
+    official: true, region: 'Europe',
+    text: 'Tangled Slinkys.'
+  },
+  {
+    expansion: 'Base',
+    id: 212,
+    official: true, region: 'Europe',
+    text: 'A moment of silence.'
+  },
+  {
+    expansion: 'Base',
+    id: 213,
+    official: true, region: 'Europe',
+    text: 'Civilian casualties.'
+  },
+  {
+    expansion: 'Base',
+    id: 214,
+    official: true, region: 'Europe',
+    text: 'Catapults.'
+  },
+  {
+    expansion: 'Base',
+    id: 215,
+    official: true, region: 'Europe',
+    text: 'Sharing needles.'
+  },
+  {
+    expansion: 'Base',
+    id: 216,
+    official: true, region: 'Europe',
+    text: 'Ethnic cleansing.'
+  },
+  {
+    expansion: 'Base',
+    id: 217,
+    official: true, region: 'Europe',
+    text: 'Emotions.'
+  },
+  {
+    expansion: 'Base',
+    id: 218,
+    official: true, region: 'Europe',
+    text: 'Children on leashes.'
+  },
+  {
+    expansion: 'Base',
+    id: 219,
+    official: true, region: 'Europe',
+    text: 'Balls.'
+  },
+  {
+    expansion: 'Base',
+    id: 220,
+    official: true, region: 'Europe',
+    text: 'Homeless people.'
+  },
+  {
+    expansion: 'Base',
+    id: 221,
+    official: true, region: 'Europe',
+    text: 'Eating all of the cookies before the AIDS bake-sale.'
+  },
+  {
+    expansion: 'Base',
+    id: 222,
+    official: true, region: 'Europe',
+    text: 'Old-people smell.'
+  },
+  {
+    expansion: 'Base',
+    id: 224,
+    official: true, region: 'Europe',
+    text: 'The inevitable heat death of the universe.'
+  },
+  {
+    expansion: 'Base',
+    id: 225,
+    official: true, region: 'Europe',
+    text: 'The violation of our most basic human rights.'
+  },
+  {
+    expansion: 'Base',
+    id: 227,
+    official: true, region: 'Europe',
+    text: 'The placenta.'
+  },
+  {
+    expansion: 'Base',
+    id: 228,
+    official: true, region: 'Europe',
+    text: 'The Rev. Dr. Martin Luther King, Jr.'
+  },
+  {
+    expansion: 'Base',
+    id: 229,
+    official: true, region: 'Europe',
+    text: 'Leprosy.'
+  },
+  {
+    expansion: 'Base',
+    id: 231,
+    official: true, region: 'Europe',
+    text: 'Multiple stab wounds.'
+  },
+  {
+    expansion: 'Base',
+    id: 232,
+    official: true, region: 'Europe',
+    text: 'Flightless birds.'
+  },
+  {
+    expansion: 'Base',
+    id: 233,
+    official: true, region: 'Europe',
+    text: 'Grave robbing.'
+  },
+  {
+    expansion: 'Base',
+    id: 234,
+    official: true, region: 'Europe',
+    text: 'Home video of Oprah sobbing into a Lean Cuisine®.'
+  },
+  {
+    expansion: 'Base',
+    id: 235,
+    official: true, region: 'Europe',
+    text: 'Oompa-Loompas.'
+  },
+  {
+    expansion: 'Base',
+    id: 236,
+    official: true, region: 'Europe',
+    text: 'A murder most foul.'
+  },
+  {
+    expansion: 'Base',
+    id: 239,
+    official: true, region: 'Europe',
+    text: 'Bill Nye the Science Guy.'
+  },
+  {
+    expansion: 'Base',
+    id: 240,
+    official: true, region: 'Europe',
+    text: 'Peeing a little bit.'
+  },
+  {
+    expansion: 'Base',
+    id: 241,
+    official: true, region: 'Europe',
+    text: 'The miracle of childbirth.'
+  },
+  {
+    expansion: 'Base',
+    id: 242,
+    official: true, region: 'Europe',
+    text: 'Tweeting.'
+  },
+  {
+    expansion: 'Base',
+    id: 244,
+    official: true, region: 'Europe',
+    text: 'Britney Spears at 55.'
+  },
+  {
+    expansion: 'Base',
+    id: 245,
+    official: true, region: 'Europe',
+    text: 'New Age music.'
+  },
+  {
+    expansion: 'Base',
+    id: 246,
+    official: true, region: 'Europe',
+    text: 'The Make-A-Wish® Foundation.'
+  },
+  {
+    expansion: 'Base',
+    id: 247,
+    official: true, region: 'Europe',
+    text: 'Firing a rifle into the air while balls deep in a squealing hog.'
+  },
+  {
+    expansion: 'Base',
+    id: 248,
+    official: true, region: 'Europe',
+    text: 'Active listening.'
+  },
+  {
+    expansion: 'Base',
+    id: 249,
+    official: true, region: 'Europe',
+    text: 'Nicolas Cage.'
+  },
+  {
+    expansion: 'Base',
+    id: 251,
+    official: true, region: 'Europe',
+    text: 'Stranger danger.'
+  },
+  {
+    expansion: 'Base',
+    id: 252,
+    official: true, region: 'Europe',
+    text: 'Hope.'
+  },
+  {
+    expansion: 'Base',
+    id: 253,
+    official: true, region: 'Europe',
+    text: 'A gassy antelope.'
+  },
+  {
+    expansion: 'Base',
+    id: 254,
+    official: true, region: 'Europe',
+    text: 'BATMAN!!!'
+  },
+  {
+    expansion: 'Base',
+    id: 255,
+    official: true, region: 'Europe',
+    text: 'Chivalry.'
+  },
+  {
+    expansion: 'Base',
+    id: 256,
+    official: true, region: 'Europe',
+    text: 'Passing a kidney stone.'
+  },
+  {
+    expansion: 'Base',
+    id: 257,
+    official: true, region: 'Europe',
+    text: 'Black people.'
+  },
+  {
+    expansion: 'Base',
+    id: 258,
+    official: true, region: 'Europe',
+    text: 'Natalie Portman.'
+  },
+  {
+    expansion: 'Base',
+    id: 259,
+    official: true, region: 'Europe',
+    text: 'A mime having a stroke.'
+  },
+  {
+    expansion: 'Base',
+    id: 260,
+    official: true, region: 'Europe',
+    text: 'Classist undertones.'
+  },
+  {
+    expansion: 'Base',
+    id: 261,
+    official: true, region: 'Europe',
+    text: 'Sean Penn.'
+  },
+  {
+    expansion: 'Base',
+    id: 262,
+    official: true, region: 'Europe',
+    text: 'A mating display.'
+  },
+  {
+    expansion: 'Base',
+    id: 263,
+    official: true, region: 'Europe',
+    text: 'The Holy Bible.'
+  },
+  {
+    expansion: 'Base',
+    id: 264,
+    official: true, region: 'Europe',
+    text: 'Hot Pockets®.'
+  },
+  {
+    expansion: 'Base',
+    id: 266,
+    official: true, region: 'Europe',
+    text: 'Pulling out.'
+  },
+  {
+    expansion: 'Base',
+    id: 267,
+    official: true, region: 'Europe',
+    text: 'Serfdom.'
+  },
+  {
+    expansion: 'Base',
+    id: 272,
+    official: true, region: 'Europe',
+    text: 'Euphoria™ by Calvin Klein.'
+  },
+  {
+    expansion: 'Base',
+    id: 273,
+    official: true, region: 'Europe',
+    text: 'The World of Warcraft.'
+  },
+  {
+    expansion: 'Base',
+    id: 275,
+    official: true, region: 'Europe',
+    text: 'The Kool-Aid Man.'
+  },
+  {
+    expansion: 'Base',
+    id: 277,
+    official: true, region: 'Europe',
+    text: 'Self-loathing.'
+  },
+  {
+    expansion: 'Base',
+    id: 278,
+    official: true, region: 'Europe',
+    text: 'A falcon with a cap on its head.'
+  },
+  {
+    expansion: 'Base',
+    id: 279,
+    official: true, region: 'Europe',
+    text: 'Historically black colleges.'
+  },
+  {
+    expansion: 'Base',
+    id: 281,
+    official: true, region: 'Europe',
+    text: 'Global warming.'
+  },
+  {
+    expansion: 'Base',
+    id: 283,
+    official: true, region: 'Europe',
+    text: 'World peace.'
+  },
+  {
+    expansion: 'Base',
+    id: 286,
+    official: true, region: 'Europe',
+    text: 'A zesty breakfast burrito.'
+  },
+  {
+    expansion: 'Base',
+    id: 287,
+    official: true, region: 'Europe',
+    text: 'Switching to Geico®.'
+  },
+  {
+    expansion: 'Base',
+    id: 288,
+    official: true, region: 'Europe',
+    text: 'Aaron Burr.'
+  },
+  {
+    expansion: 'Base',
+    id: 290,
+    official: true, region: 'Europe',
+    text: 'Land mines.'
+  },
+  {
+    expansion: 'Base',
+    id: 291,
+    official: true, region: 'Europe',
+    text: 'Former President George W. Bush.'
+  },
+  {
+    expansion: 'Base',
+    id: 292,
+    official: true, region: 'Europe',
+    text: 'Geese.'
+  },
+  {
+    expansion: 'Base',
+    id: 293,
+    official: true, region: 'Europe',
+    text: 'Mutually-assured destruction.'
+  },
+  {
+    expansion: 'Base',
+    id: 294,
+    official: true, region: 'Europe',
+    text: 'College.'
+  },
+  {
+    expansion: 'Base',
+    id: 296,
+    official: true, region: 'Europe',
+    text: 'Bling.'
+  },
+  {
+    expansion: 'Base',
+    id: 298,
+    official: true, region: 'Europe',
+    text: 'A time travel paradox.'
+  },
+  {
+    expansion: 'Base',
+    id: 299,
+    official: true, region: 'Europe',
+    text: 'Seppuku.'
+  },
+  {
+    expansion: 'Base',
+    id: 300,
+    official: true, region: 'Europe',
+    text: 'Poor life choices.'
+  },
+  {
+    expansion: 'Base',
+    id: 302,
+    official: true, region: 'Europe',
+    text: 'Italians.'
+  },
+  {
+    expansion: 'Base',
+    id: 303,
+    official: true, region: 'Europe',
+    text: 'GoGurt®.'
+  },
+  {
+    expansion: 'Base',
+    id: 304,
+    official: true, region: 'Europe',
+    text: 'Finger painting.'
+  },
+  {
+    expansion: 'Base',
+    id: 305,
+    official: true, region: 'Europe',
+    text: 'Robert Downey, Jr.'
+  },
+  {
+    expansion: 'Base',
+    id: 307,
+    official: true, region: 'Europe',
+    text: 'Smegma.'
+  },
+  {
+    expansion: 'Base',
+    id: 309,
+    official: true, region: 'Europe',
+    text: 'The South.'
+  },
+  {
+    expansion: 'Base',
+    id: 310,
+    official: true, region: 'Europe',
+    text: 'Christopher Walken.'
+  },
+  {
+    expansion: 'Base',
+    id: 311,
+    official: true, region: 'Europe',
+    text: 'Gloryholes.'
+  },
+  {
+    expansion: 'Base',
+    id: 313,
+    official: true, region: 'Europe',
+    text: 'Public ridicule.'
+  },
+  {
+    expansion: 'Base',
+    id: 314,
+    official: true, region: 'Europe',
+    text: 'A tiny horse.'
+  },
+  {
+    expansion: 'Base',
+    id: 315,
+    official: true, region: 'Europe',
+    text: 'Arnold Schwarzenegger.'
+  },
+  {
+    expansion: 'Base',
+    id: 316,
+    official: true, region: 'Europe',
+    text: 'A stray pube.'
+  },
+  {
+    expansion: 'Base',
+    id: 318,
+    official: true, region: 'Europe',
+    text: 'Child abuse.'
+  },
+  {
+    expansion: 'Base',
+    id: 322,
+    official: true, region: 'Europe',
+    text: 'Re-gifting.'
+  },
+  {
+    expansion: 'Base',
+    id: 324,
+    official: true, region: 'Europe',
+    text: 'A sausage festival.'
+  },
+  {
+    expansion: 'Base',
+    id: 326,
+    official: true, region: 'Europe',
+    text: 'Drinking alone.'
+  },
+  {
+    expansion: 'Base',
+    id: 327,
+    official: true, region: 'Europe',
+    text: 'Too much hair gel.'
+  },
+  {
+    expansion: 'Base',
+    id: 328,
+    official: true, region: 'Europe',
+    text: 'Hulk Hogan.'
+  },
+  {
+    expansion: 'Base',
+    id: 329,
+    official: true, region: 'Europe',
+    text: 'Overcompensation.'
+  },
+  {
+    expansion: 'Base',
+    id: 330,
+    official: true, region: 'Europe',
+    text: 'Foreskin.'
+  },
+  {
+    expansion: 'Base',
+    id: 331,
+    official: true, region: 'Europe',
+    text: 'Free samples.'
+  },
+  {
+    expansion: 'Base',
+    id: 332,
+    official: true, region: 'Europe',
+    text: "Shaquille O'Neal's acting career."
+  },
+  {
+    expansion: 'Base',
+    id: 333,
+    official: true, region: 'Europe',
+    text: 'Five-Dollar Footlongs™.'
+  },
+  {
+    expansion: 'Base',
+    id: 334,
+    official: true, region: 'Europe',
+    text: 'Whipping it out.'
+  },
+  {
+    expansion: 'Base',
+    id: 338,
+    official: true, region: 'Europe',
+    text: 'Dental dams.'
+  },
+  {
+    expansion: 'Base',
+    id: 339,
+    official: true, region: 'Europe',
+    text: 'Being a dick to children.'
+  },
+  {
+    expansion: 'Base',
+    id: 340,
+    official: true, region: 'Europe',
+    text: 'Famine.'
+  },
+  {
+    expansion: 'Base',
+    id: 341,
+    official: true, region: 'Europe',
+    text: 'Chainsaws for hands.'
+  },
+  {
+    expansion: 'Base',
+    id: 342,
+    official: true, region: 'Europe',
+    text: 'A gypsy curse.'
+  },
+  {
+    expansion: 'Base',
+    id: 343,
+    official: true, region: 'Europe',
+    text: 'AXE Body Spray.'
+  },
+  {
+    expansion: 'Base',
+    id: 344,
+    official: true, region: 'Europe',
+    text: 'The Force.'
+  },
+  {
+    expansion: 'Base',
+    id: 345,
+    official: true, region: 'Europe',
+    text: 'Explosions.'
+  },
+  {
+    expansion: 'Base',
+    id: 346,
+    official: true, region: 'Europe',
+    text: 'Cybernetic enhancements.'
+  },
+  {
+    expansion: 'Base',
+    id: 347,
+    official: true, region: 'Europe',
+    text: 'Customer service representatives.'
+  },
+  {
+    expansion: 'Base',
+    id: 348,
+    official: true, region: 'Europe',
+    text: 'White privilege.'
+  },
+  {
+    expansion: 'Base',
+    id: 349,
+    official: true, region: 'Europe',
+    text: 'Gandhi.'
+  },
+  {
+    expansion: 'Base',
+    id: 350,
+    official: true, region: 'Europe',
+    text: 'Road head.'
+  },
+  {
+    expansion: 'Base',
+    id: 351,
+    official: true, region: 'Europe',
+    text: 'God.'
+  },
+  {
+    expansion: 'Base',
+    id: 352,
+    official: true, region: 'Europe',
+    text: 'Poorly-timed Holocaust jokes.'
+  },
+  {
+    expansion: 'Base',
+    id: 353,
+    official: true, region: 'Europe',
+    text: '8 oz. of sweet Mexican black-tar heroin.'
+  },
+  {
+    expansion: 'Base',
+    id: 354,
+    official: true, region: 'Europe',
+    text: 'Judge Judy.'
+  },
+  {
+    expansion: 'Base',
+    id: 355,
+    official: true, region: 'Europe',
+    text: 'The Little Engine That Could.'
+  },
+  {
+    expansion: 'Base',
+    id: 356,
+    official: true, region: 'Europe',
+    text: 'Altar boys.'
+  },
+  {
+    expansion: 'Base',
+    id: 357,
+    official: true, region: 'Europe',
+    text: 'Mr. Clean, right behind you.'
+  },
+  {
+    expansion: 'Base',
+    id: 359,
+    official: true, region: 'Europe',
+    text: 'Dwarf tossing.'
+  },
+  {
+    expansion: 'Base',
+    id: 360,
+    official: true, region: 'Europe',
+    text: 'Friction.'
+  },
+  {
+    expansion: 'Base',
+    id: 361,
+    official: true, region: 'Europe',
+    text: 'Lady Gaga.'
+  },
+  {
+    expansion: 'Base',
+    id: 362,
+    official: true, region: 'Europe',
+    text: 'Scientology.'
+  },
+  {
+    expansion: 'Base',
+    id: 363,
+    official: true, region: 'Europe',
+    text: 'Justin Bieber.'
+  },
+  {
+    expansion: 'Base',
+    id: 364,
+    official: true, region: 'Europe',
+    text: 'A death ray.'
+  },
+  {
+    expansion: 'Base',
+    id: 365,
+    official: true, region: 'Europe',
+    text: 'Vigilante justice.'
+  },
+  {
+    expansion: 'Base',
+    id: 366,
+    official: true, region: 'Europe',
+    text: 'The Pope.'
+  },
+  {
+    expansion: 'Base',
+    id: 367,
+    official: true, region: 'Europe',
+    text: 'A sea of troubles.'
+  },
+  {
+    expansion: 'Base',
+    id: 369,
+    official: true, region: 'Europe',
+    text: 'Poor people.'
+  },
+  {
+    expansion: 'Base',
+    id: 370,
+    official: true, region: 'Europe',
+    text: 'A fetus.'
+  },
+  {
+    expansion: 'Base',
+    id: 372,
+    official: true, region: 'Europe',
+    text: "Exactly what you'd expect."
+  },
+  {
+    expansion: 'Base',
+    id: 375,
+    official: true, region: 'Europe',
+    text: 'A balanced breakfast.'
+  },
+  {
+    expansion: 'Base',
+    id: 377,
+    official: true, region: 'Europe',
+    text: 'Lockjaw.'
+  },
+  {
+    expansion: 'Base',
+    id: 379,
+    official: true, region: 'Europe',
+    text: 'Take-backsies.'
+  },
+  {
+    expansion: 'Base',
+    id: 381,
+    official: true, region: 'Europe',
+    text: 'Opposable thumbs.'
+  },
+  {
+    expansion: 'Base',
+    id: 384,
+    official: true, region: 'Europe',
+    text: 'Inappropriate yodeling.'
+  },
+  {
+    expansion: 'Base',
+    id: 385,
+    official: true, region: 'Europe',
+    text: 'Golden showers.'
+  },
+  {
+    expansion: 'Base',
+    id: 387,
+    official: true, region: 'Europe',
+    text: 'Copping a feel.'
+  },
+  {
+    expansion: 'Base',
+    id: 388,
+    official: true, region: 'Europe',
+    text: 'Auschwitz.'
+  },
+  {
+    expansion: 'Base',
+    id: 389,
+    official: true, region: 'Europe',
+    text: 'Elderly Japanese men.'
+  },
+  {
+    expansion: 'Base',
+    id: 396,
+    official: true, region: 'Europe',
+    text: 'Hurricane Katrina.'
+  },
+  {
+    expansion: 'Base',
+    id: 398,
+    official: true, region: 'Europe',
+    text: 'Science.'
+  },
+  {
+    expansion: 'Base',
+    id: 400,
+    official: true, region: 'Europe',
+    text: 'Cards Against Humanity.'
+  },
+  {
+    expansion: 'Base',
+    id: 401,
+    official: true, region: 'Europe',
+    text: 'Fancy Feast®.'
+  },
+  {
+    expansion: 'Base',
+    id: 403,
+    official: true, region: 'Europe',
+    text: 'Lumberjack fantasies.'
+  },
+  {
+    expansion: 'Base',
+    id: 404,
+    official: true, region: 'Europe',
+    text: 'American Gladiators.'
+  },
+  {
+    expansion: 'Base',
+    id: 406,
+    official: true, region: 'Europe',
+    text: 'Sean Connery.'
+  },
+  {
+    expansion: 'Base',
+    id: 407,
+    official: true, region: 'Europe',
+    text: 'William Shatner.'
+  },
+  {
+    expansion: 'Base',
+    id: 408,
+    official: true, region: 'Europe',
+    text: "Domino's™ Oreo™ Dessert Pizza."
+  },
+  {
+    expansion: 'Base',
+    id: 410,
+    official: true, region: 'Europe',
+    text: 'Centaurs.'
+  },
+  {
+    expansion: 'Base',
+    id: 413,
+    official: true, region: 'Europe',
+    text: 'The milk man.'
+  },
+  {
+    expansion: 'Base',
+    id: 414,
+    official: true, region: 'Europe',
+    text: 'Waterboarding.'
+  },
+  {
+    expansion: 'Base',
+    id: 415,
+    official: true, region: 'Europe',
+    text: 'Wifely duties.'
+  },
+  {
+    expansion: 'Base',
+    id: 416,
+    official: true, region: 'Europe',
+    text: 'Loose lips.'
+  },
+  {
+    expansion: 'Base',
+    id: 417,
+    official: true, region: 'Europe',
+    text: 'The Blood of Christ.'
+  },
+  {
+    expansion: 'Base',
+    id: 418,
+    official: true, region: 'Europe',
+    text: 'Actually taking candy from a baby.'
+  },
+  {
+    expansion: 'Base',
+    id: 419,
+    official: true, region: 'Europe',
+    text: 'The token minority.'
+  },
+  {
+    expansion: 'Base',
+    id: 420,
+    official: true, region: 'Europe',
+    text: 'Jibber-jabber.'
+  },
+  {
+    expansion: 'Base',
+    id: 422,
+    official: true, region: 'Europe',
+    text: 'Bingeing and purging.'
+  },
+  {
+    expansion: 'Base',
+    id: 423,
+    official: true, region: 'Europe',
+    text: 'A clandestine butt scratch.'
+  },
+  {
+    expansion: 'Base',
+    id: 424,
+    official: true, region: 'Europe',
+    text: 'The Chinese gymnastics team.'
+  },
+  {
+    expansion: 'Base',
+    id: 426,
+    official: true, region: 'Europe',
+    text: 'The Hamburglar.'
+  },
+  {
+    expansion: 'Base',
+    id: 427,
+    official: true, region: 'Europe',
+    text: 'Police brutality.'
+  },
+  {
+    expansion: 'Base',
+    id: 428,
+    official: true, region: 'Europe',
+    text: 'Man meat.'
+  },
+  {
+    expansion: 'Base',
+    id: 429,
+    official: true, region: 'Europe',
+    text: 'Forgetting the Alamo.'
+  },
+  {
+    expansion: 'Base',
+    id: 431,
+    official: true, region: 'Europe',
+    text: 'Crystal meth.'
+  },
+  {
+    expansion: 'Base',
+    id: 434,
+    official: true, region: 'Europe',
+    text: 'Third base.'
+  },
+  {
+    expansion: 'Base',
+    id: 435,
+    official: true, region: 'America',
+    text: 'Soiling oneself.'
+  },
+  {
+    expansion: 'Base',
+    id: 436,
+    official: true, region: 'America',
+    text: 'Laying an egg.'
+  },
+  {
+    expansion: 'Base',
+    id: 437,
+    official: true, region: 'America',
+    text: 'Giving 110%.'
+  },
+  {
+    expansion: 'Base',
+    id: 438,
+    official: true, region: 'America',
+    text: 'Hot people.'
+  },
+  {
+    expansion: 'Base',
+    id: 439,
+    official: true, region: 'America',
+    text: 'Friendly fire.'
+  },
+  {
+    expansion: 'Base',
+    id: 440,
+    official: true, region: 'America',
+    text: 'Count Chocula.'
+  },
+  {
+    expansion: 'Base',
+    id: 441,
+    official: true, region: 'America',
+    text: 'Pac-Man uncontrollably guzzling cum.'
+  },
+  {
+    expansion: 'Base',
+    id: 442,
+    official: true, region: 'America',
+    text: 'Estrogen.'
+  },
+  {
+    expansion: 'Base',
+    id: 444,
+    official: true, region: 'America',
+    text: 'Kanye West.'
+  },
+  {
+    expansion: 'Base',
+    id: 445,
+    official: true, region: 'America',
+    text: 'A robust mongoloid.'
+  },
+  {
+    expansion: 'Base',
+    id: 446,
+    official: true, region: 'America',
+    text: 'The Donald Trump Seal of Approval™.'
+  },
+  {
+    expansion: 'Base',
+    id: 447,
+    official: true, region: 'America',
+    text: 'The true meaning of Christmas.'
+  },
+  {
+    expansion: 'Base',
+    id: 448,
+    official: true, region: 'America',
+    text: 'Her Royal Highness, Queen Elizabeth II.'
+  },
+  {
+    expansion: 'Base',
+    id: 449,
+    official: true, region: 'America',
+    text: 'An honest cop with nothing left to lose.'
+  },
+  {
+    expansion: 'Base',
+    id: 450,
+    official: true, region: 'America',
+    text: "Feeding Rosie O'Donnell."
+  },
+  {
+    expansion: 'Base',
+    id: 451,
+    official: true, region: 'America',
+    text: 'The Amish.'
+  },
+  {
+    expansion: 'Base',
+    id: 452,
+    official: true, region: 'America',
+    text: 'The terrorists.'
+  },
+  {
+    expansion: 'Base',
+    id: 454,
+    official: true, region: 'America',
+    text: 'Pooping back and forth. Forever.'
+  },
+  {
+    expansion: 'Base',
+    id: 455,
+    official: true, region: 'America',
+    text: 'Friends who eat all the snacks.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 460,
+    official: true, region: 'America',
+    text: 'A beached whale.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 461,
+    official: true, region: 'America',
+    text: 'A bloody pacifier.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 462,
+    official: true, region: 'America',
+    text: 'A crappy little hand.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 463,
+    official: true, region: 'America',
+    text: 'A low standard of living.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 464,
+    official: true, region: 'America',
+    text: 'A nuanced critique.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 465,
+    official: true, region: 'America',
+    text: 'Panty raids.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 466,
+    official: true, region: 'America',
+    text: 'A passionate Latino lover.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 467,
+    official: true, region: 'America',
+    text: 'A rival dojo.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 468,
+    official: true, region: 'America',
+    text: 'A web of lies.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 470,
+    official: true, region: 'America',
+    text: 'Clams.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 471,
+    official: true, region: 'America',
+    text: 'Apologizing.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 473,
+    official: true, region: 'America',
+    text: 'Neil Patrick Harris.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 475,
+    official: true, region: 'America',
+    text: 'Being a dinosaur.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 477,
+    official: true, region: 'America',
+    text: 'Bosnian chicken farmers.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 478,
+    official: true, region: 'America',
+    text: 'Nubile slave boys.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 479,
+    official: true, region: 'America',
+    text: 'Carnies.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 484,
+    official: true, region: 'America',
+    text: 'Dorito breath.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 485,
+    official: true, region: 'America',
+    text: 'Eating an albino.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 486,
+    official: true, region: 'America',
+    text: 'Enormous Scandinavian women.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 487,
+    official: true, region: 'America',
+    text: 'Fabricating statistics.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 488,
+    official: true, region: 'America',
+    text: 'Finding a skeleton.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 489,
+    official: true, region: 'America',
+    text: 'Gandalf.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 490,
+    official: true, region: 'America',
+    text: 'Genetically engineered super-soldiers.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 491,
+    official: true, region: 'America',
+    text: "George Clooney's musk."
+  },
+  {
+    expansion: 'CAHe1',
+    id: 496,
+    official: true, region: 'America',
+    text: 'Hipsters.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 497,
+    official: true, region: 'America',
+    text: 'Historical revisionism.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 498,
+    official: true, region: 'America',
+    text: 'Insatiable bloodlust.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 499,
+    official: true, region: 'America',
+    text: 'Jafar.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 500,
+    official: true, region: 'America',
+    text: 'Jean-Claude Van Damme.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 501,
+    official: true, region: 'America',
+    text: 'Just the tip.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 502,
+    official: true, region: 'America',
+    text: 'Mad hacky-sack skills.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 503,
+    official: true, region: 'America',
+    text: 'Leveling up.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 506,
+    official: true, region: 'America',
+    text: '24-hour media coverage.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 507,
+    official: true, region: 'America',
+    text: 'Medieval Times© Dinner \u0026 Tournament.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 509,
+    official: true, region: 'America',
+    text: 'My machete.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 510,
+    official: true, region: 'America',
+    text: 'One thousand Slim Jims.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 511,
+    official: true, region: 'America',
+    text: 'Ominous background music.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 512,
+    official: true, region: 'America',
+    text: 'Overpowering your father.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 513,
+    official: true, region: 'America',
+    text: 'Stockholm Syndrome.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 514,
+    official: true, region: 'America',
+    text: 'Quiche.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 515,
+    official: true, region: 'America',
+    text: 'Quivering jowls.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 518,
+    official: true, region: 'America',
+    text: 'Ryan Gosling riding in on a white horse.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 519,
+    official: true, region: 'America',
+    text: 'Santa Claus.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 520,
+    official: true, region: 'America',
+    text: 'Scrotum tickling.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 523,
+    official: true, region: 'America',
+    text: 'Saliva.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 524,
+    official: true, region: 'America',
+    text: 'Space muffins.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 525,
+    official: true, region: 'America',
+    text: 'Statistically validated stereotypes.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 527,
+    official: true, region: 'America',
+    text: 'The boners of the elderly.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 528,
+    official: true, region: 'America',
+    text: 'The economy.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 529,
+    official: true, region: 'America',
+    text: 'Syphilitic insanity.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 530,
+    official: true, region: 'America',
+    text: 'The Gulags.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 531,
+    official: true, region: 'America',
+    text: 'The harsh light of day.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 532,
+    official: true, region: 'America',
+    text: 'The hiccups.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 533,
+    official: true, region: 'America',
+    text: 'The shambling corpse of Larry King.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 534,
+    official: true, region: 'America',
+    text: 'The four arms of Vishnu.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 535,
+    official: true, region: 'America',
+    text: 'Being a busy adult with many important things to do.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 536,
+    official: true, region: 'America',
+    text: 'Tripping balls.'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 537,
+    official: true, region: 'America',
+    text: 'Words, words, words.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 540,
+    official: true, region: 'America',
+    text: "The mere concept of Applebee's®."
+  },
+  {
+    expansion: 'CAHe2',
+    id: 542,
+    official: true, region: 'America',
+    text: 'Catastrophic urethral trauma.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 543,
+    official: true, region: 'America',
+    text: "Hillary Clinton's death stare."
+  },
+  {
+    expansion: 'CAHe2',
+    id: 544,
+    official: true, region: 'America',
+    text: 'Existing.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 546,
+    official: true, region: 'America',
+    text: 'Mooing.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 548,
+    official: true, region: 'America',
+    text: "Daddy's belt."
+  },
+  {
+    expansion: 'CAHe2',
+    id: 550,
+    official: true, region: 'America',
+    text: 'Weapons-grade plutonium.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 553,
+    official: true, region: 'America',
+    text: 'Rising from the grave.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 554,
+    official: true, region: 'America',
+    text: 'The mixing of the races.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 556,
+    official: true, region: 'America',
+    text: 'Scrotal frostbite.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 557,
+    official: true, region: 'America',
+    text: 'All of this blood.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 558,
+    official: true, region: 'America',
+    text: 'Loki, the trickster god.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 561,
+    official: true, region: 'America',
+    text: 'Tongue.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 562,
+    official: true, region: 'America',
+    text: 'Finding Waldo.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 563,
+    official: true, region: 'America',
+    text: 'Upgrading homeless people to mobile hotspots.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 564,
+    official: true, region: 'America',
+    text: 'Wearing an octopus for a hat.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 565,
+    official: true, region: 'America',
+    text: 'An unhinged ferris wheel rolling toward the sea.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 566,
+    official: true, region: 'America',
+    text: 'Living in a trashcan.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 567,
+    official: true, region: 'America',
+    text: 'The corporations.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 568,
+    official: true, region: 'America',
+    text: 'A magic hippie love cloud.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 570,
+    official: true, region: 'America',
+    text: "Survivor's guilt."
+  },
+  {
+    expansion: 'CAHe2',
+    id: 571,
+    official: true, region: 'America',
+    text: 'Me.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 572,
+    official: true, region: 'America',
+    text: 'Getting hilariously gang-banged by the Blue Man Group.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 573,
+    official: true, region: 'America',
+    text: 'Jeff Goldblum.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 574,
+    official: true, region: 'America',
+    text: 'Making a friend.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 575,
+    official: true, region: 'America',
+    text: "A soulful rendition of Ol' Man River."
+  },
+  {
+    expansion: 'CAHe2',
+    id: 577,
+    official: true, region: 'America',
+    text: 'A sweaty, panting leather daddy.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 578,
+    official: true, region: 'America',
+    text: 'Spring break!'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 580,
+    official: true, region: 'America',
+    text: 'Dining with cardboard cutouts of the cast of Friends.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 582,
+    official: true, region: 'America',
+    text: "Beefin' over turf."
+  },
+  {
+    expansion: 'CAHe2',
+    id: 583,
+    official: true, region: 'America',
+    text: 'A squadron of moles wearing aviator goggles.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 584,
+    official: true, region: 'America',
+    text: 'Bullshit.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 585,
+    official: true, region: 'America',
+    text: 'The Google.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 587,
+    official: true, region: 'America',
+    text: 'The new Radiohead album.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 588,
+    official: true, region: 'America',
+    text: 'An army of skeletons.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 590,
+    official: true, region: 'America',
+    text: 'Mild autism.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 591,
+    official: true, region: 'America',
+    text: 'Nunchuck moves.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 593,
+    official: true, region: 'America',
+    text: 'An ether-soaked rag.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 594,
+    official: true, region: 'America',
+    text: 'A sweet spaceship.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 595,
+    official: true, region: 'America',
+    text: 'A 55-gallon drum of lube.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 596,
+    official: true, region: 'America',
+    text: 'Special musical guest, Cher.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 597,
+    official: true, region: 'America',
+    text: 'The human body.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 598,
+    official: true, region: 'America',
+    text: 'Boris the Soviet Love Hammer.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 599,
+    official: true, region: 'America',
+    text: 'The grey nutrient broth that sustains Mitt Romney.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 601,
+    official: true, region: 'America',
+    text: 'Power.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 602,
+    official: true, region: 'America',
+    text: 'Oncoming traffic.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 603,
+    official: true, region: 'America',
+    text: 'A dollop of sour cream.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 604,
+    official: true, region: 'America',
+    text: 'A slightly shittier parallel universe.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 608,
+    official: true, region: 'America',
+    text: 'The day the birds attacked.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 609,
+    official: true, region: 'America',
+    text: 'One Ring to rule them all.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 610,
+    official: true, region: 'America',
+    text: "Grandpa's ashes."
+  },
+  {
+    expansion: 'CAHe2',
+    id: 611,
+    official: true, region: 'America',
+    text: 'Basic human decency.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 612,
+    official: true, region: 'America',
+    text: 'A Burmese tiger pit.'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 613,
+    official: true, region: 'America',
+    text: 'Death by Steven Seagal'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 786,
+    official: true, region: 'America',
+    text: "Santa's heavy sack."
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 787,
+    official: true, region: 'America',
+    text: 'Clearing a bloody path through Walmart with a scimitar.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 788,
+    official: true, region: 'America',
+    text: 'Another shitty year.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 789,
+    official: true, region: 'America',
+    text: 'Whatever Kwanzaa is supposed to be about.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 790,
+    official: true, region: 'America',
+    text: 'A Christmas stocking full of coleslaw.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 791,
+    official: true, region: 'America',
+    text: 'Elf cum.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 792,
+    official: true, region: 'America',
+    text: 'The tiny, calloused hands of the Chinese children that made this card.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 794,
+    official: true, region: 'America',
+    text: 'Socks.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 795,
+    official: true, region: 'America',
+    text: 'Pretending to be happy.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 796,
+    official: true, region: 'America',
+    text: 'Krampus, the Austrian Christmas monster.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 797,
+    official: true, region: 'America',
+    text: 'The Star Wars Holiday Special.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 798,
+    official: true, region: 'America',
+    text: 'My hot cousin.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 799,
+    official: true, region: 'America',
+    text: 'Mall Santa.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 800,
+    official: true, region: 'America',
+    text: 'Several intertwining love stories featuring Hugh Grant.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 801,
+    official: true, region: 'America',
+    text: 'A Hungry-Man™ Frozen Christmas Dinner for one.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 802,
+    official: true, region: 'America',
+    text: 'Gift-wrapping a live hamster.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 803,
+    official: true, region: 'America',
+    text: 'Space Jam on VHS.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 804,
+    official: true, region: 'America',
+    text: 'Immaculate conception.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 806,
+    official: true, region: 'America',
+    text: 'A visually arresting turtleneck.'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 807,
+    official: true, region: 'America',
+    text: 'A toxic family environment.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 943,
+    official: true, region: 'America',
+    text: 'Gay aliens.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 946,
+    official: true, region: 'America',
+    text: "The Quesadilla Explosion Salad™ from Chili's©."
+  },
+  {
+    expansion: 'CAHe3',
+    id: 947,
+    official: true, region: 'America',
+    text: 'Actually getting shot, for real.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 951,
+    official: true, region: 'America',
+    text: 'Nothing.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 954,
+    official: true, region: 'America',
+    text: 'The systematic destruction of an entire people and their way of life.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 955,
+    official: true, region: 'America',
+    text: 'Samuel L. Jackson.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 958,
+    official: true, region: 'America',
+    text: 'The entire Internet.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 966,
+    official: true, region: 'America',
+    text: 'Putting an entire peanut butter and jelly sandwich into the VCR.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 967,
+    official: true, region: 'America',
+    text: 'Spending lots of money.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 970,
+    official: true, region: 'America',
+    text: 'A greased-up Matthew McConaughey.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 971,
+    official: true, region: 'America',
+    text: 'An unstoppable wave of fire ants.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 972,
+    official: true, region: 'America',
+    text: 'Not contributing to society in any meaningful way.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 973,
+    official: true, region: 'America',
+    text: "An all-midget production of Shakespeare's \u003ci\u003eRichard III\u003c/i\u003e."
+  },
+  {
+    expansion: 'CAHe3',
+    id: 975,
+    official: true, region: 'America',
+    text: 'The moist, demanding chasm of his mouth.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 976,
+    official: true, region: 'America',
+    text: 'Filling every orifice with butterscotch pudding.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 977,
+    official: true, region: 'America',
+    text: 'Unlimited soup, salad, and breadsticks.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 978,
+    official: true, region: 'America',
+    text: 'Crying into the pages of Sylvia Plath.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 979,
+    official: true, region: 'America',
+    text: 'Velcro™.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 980,
+    official: true, region: 'America',
+    text: 'A PowerPoint presentation.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 981,
+    official: true, region: 'America',
+    text: 'A surprising amount of hair.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 982,
+    official: true, region: 'America',
+    text: "Eating Tom Selleck's mustache to gain his powers."
+  },
+  {
+    expansion: 'CAHe3',
+    id: 990,
+    official: true, region: 'America',
+    text: 'The Land of Chocolate.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 991,
+    official: true, region: 'America',
+    text: 'Slapping a racist old lady.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 992,
+    official: true, region: 'America',
+    text: 'A lamprey swimming up the toilet and latching onto your taint.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 993,
+    official: true, region: 'America',
+    text: 'Jumping out at people.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 994,
+    official: true, region: 'America',
+    text: 'A black male in his early 20s, last seen wearing a hoodie.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 998,
+    official: true, region: 'America',
+    text: 'The Harlem Globetrotters.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 1000,
+    official: true, region: 'America',
+    text: 'My manservant, Claude.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 1002,
+    official: true, region: 'America',
+    text: 'Letting everyone down.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 1003,
+    official: true, region: 'America',
+    text: 'A spontaneous conga line.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 1005,
+    official: true, region: 'America',
+    text: 'Disco fever.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 1009,
+    official: true, region: 'America',
+    text: "Girls that always be textin'."
+  },
+  {
+    expansion: 'CAHe3',
+    id: 1010,
+    official: true, region: 'America',
+    text: 'Blowing some dudes in an alley.'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 1011,
+    official: true, region: 'America',
+    text: 'Drinking ten 5-hour ENERGYs® to get fifty continuous hours of energy.'
+  }
+];
+
+module.exports = answers;
