+++ seed_data/questions-data.js
@@ -1,0 +1,1303 @@
+/* eslint-disable */
+const questions = [
+  {
+    expansion: 'Base',
+    id: 1,
+    numAnswers: 1,
+    official: true,
+    text: "_?  There's an app for that.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 2,
+    numAnswers: 1,
+    official: true,
+    text: "Why can't I sleep at night?",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 3,
+    numAnswers: 1,
+    official: true,
+    text: "What's that smell?",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 4,
+    numAnswers: 1,
+    official: true,
+    text: "I got 99 problems but _ ain't one.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 5,
+    numAnswers: 1,
+    official: true,
+    text: "Maybe she's born with it.  Maybe it's _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 6,
+    numAnswers: 1,
+    official: true,
+    text: "What's the next Happy Meal® toy?",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 7,
+    numAnswers: 1,
+    official: true,
+    text: 'Anthropologists have recently discovered a primitive tribe that worships _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 8,
+    numAnswers: 1,
+    official: true,
+    text: "It's a pity that kids these days are all getting involved with _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 9,
+    numAnswers: 1,
+    official: true,
+    text: "During Picasso's often-overlooked Brown Period, he produced hundreds of paintings of _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 10,
+    numAnswers: 1,
+    official: true,
+    text: 'Alternative medicine is now embracing the curative powers of _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 11,
+    numAnswers: 2,
+    official: true,
+    text: 'And the Academy Award for _ goes to _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 12,
+    numAnswers: 1,
+    official: true,
+    text: "What's that sound?",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 13,
+    numAnswers: 1,
+    official: true,
+    text: 'What ended my last relationship?',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 14,
+    numAnswers: 1,
+    official: true,
+    text: "MTV's new reality show features eight washed-up celebrities living with _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 15,
+    numAnswers: 1,
+    official: true,
+    text: 'I drink to forget _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 16,
+    numAnswers: 1,
+    official: true,
+    text:
+    "I'm sorry professor, but I couldn't complete my homework because of _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 17,
+    numAnswers: 1,
+    official: true,
+    text: "What is Batman's guilty pleasure?",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 18,
+    numAnswers: 1,
+    official: true,
+    text: 'This is the way the world ends \u003cbr\u003e This is the way the world ends \u003cbr\u003e Not with a bang but with _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 19,
+    numAnswers: 1,
+    official: true,
+    text: "What's a girl's best friend?",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 20,
+    numAnswers: 1,
+    official: true,
+    text: 'TSA guidelines now prohibit _ on airplanes.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 21,
+    numAnswers: 1,
+    official: true,
+    text: "_.  That's how I want to die.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 22,
+    numAnswers: 2,
+    official: true,
+    text: 'For my next trick, I will pull _ out of _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 23,
+    numAnswers: 1,
+    official: true,
+    text: 'In the new Disney Channel Original Movie, Hannah Montana struggles with _ for the first time.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 24,
+    numAnswers: 2,
+    official: true,
+    text: '_ is a slippery slope that leads to _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 25,
+    numAnswers: 1,
+    official: true,
+    text: 'What does Dick Cheney prefer?',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 26,
+    numAnswers: 1,
+    official: true,
+    text: "Dear Abby, I'm having some trouble with _ and would like your advice.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 27,
+    numAnswers: 1,
+    official: true,
+    text: 'Instead of coal, Santa now gives the bad children _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 28,
+    numAnswers: 1,
+    official: true,
+    text: "What's the most emo?",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 29,
+    numAnswers: 1,
+    official: true,
+    text: 'In 1,000 years when paper money is but a distant memory, _ will be our currency.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 30,
+    numAnswers: 2,
+    official: true,
+    text: "What's the next superhero/sidekick duo?",
+    region: 'Africa'
+  },
+  {
+    expansion: 'Base',
+    id: 31,
+    numAnswers: 2,
+    official: true,
+    text: "In M. Night Shyamalan's new movie, Bruce Willis discovers that _ had really been _ all along.",
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 32,
+    numAnswers: 1,
+    official: true,
+    text: 'A romantic, candlelit dinner would be incomplete without _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 33,
+    numAnswers: 1,
+    official: true,
+    text: "_.  Becha can't have just one!",
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 34,
+    numAnswers: 1,
+    official: true,
+    text: 'White people like _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 35,
+    numAnswers: 1,
+    official: true,
+    text: '_.  High five, bro.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 36,
+    numAnswers: 1,
+    official: true,
+    text: 'Next from J.K. Rowling: Harry Potter and the Chamber of _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 37,
+    numAnswers: 1,
+    official: true,
+    text: 'BILLY MAYS HERE FOR _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 38,
+    numAnswers: 2,
+    official: true,
+    text: 'In a world ravaged by _, our only solace is _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 39,
+    numAnswers: 1,
+    official: true,
+    text: 'War!  What is it good for?',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 40,
+    numAnswers: 1,
+    official: true,
+    text: 'During sex, I like to think about _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 41,
+    numAnswers: 1,
+    official: true,
+    text: 'What are my parents hiding from me?',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 42,
+    numAnswers: 1,
+    official: true,
+    text: 'What will always get you laid?',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 43,
+    numAnswers: 1,
+    official: true,
+    text: 'In L.A. County Jail, word is you can trade 200 cigarettes for _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 44,
+    numAnswers: 1,
+    official: true,
+    text: 'What did I bring back from Mexico?',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 45,
+    numAnswers: 1,
+    official: true,
+    text: "What don't you want to find in your Chinese food?",
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 46,
+    numAnswers: 1,
+    official: true,
+    text: 'What will I bring back in time to convince people that I am a powerful wizard?',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 47,
+    numAnswers: 1,
+    official: true,
+    text: 'How am I maintaining my relationship status?',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 48,
+    numAnswers: 1,
+    official: true,
+    text: "_.  It's a trap!",
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 49,
+    numAnswers: 1,
+    official: true,
+    text: 'Coming to Broadway this season, _: The Musical.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 50,
+    numAnswers: 1,
+    official: true,
+    text: 'While the United States raced the Soviet Union to the moon, the Mexican government funneled millions of pesos into research on _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 51,
+    numAnswers: 1,
+    official: true,
+    text: 'After the earthquake, Sean Penn brought _ to the people of Haiti.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 52,
+    numAnswers: 1,
+    official: true,
+    text: 'Next on ESPN2, the World Series of _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 53,
+    numAnswers: 2,
+    official: true,
+    text: 'Step 1: _.  Step 2: _.  Step 3: Profit.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 54,
+    numAnswers: 2,
+    official: true,
+    text: "Rumor has it that Vladimir Putin's favorite dish is _ stuffed with _.",
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 55,
+    numAnswers: 1,
+    official: true,
+    text: 'But before I kill you, Mr. Bond, I must show you _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 56,
+    numAnswers: 1,
+    official: true,
+    text: 'What gives me uncontrollable gas?',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 57,
+    numAnswers: 1,
+    official: true,
+    text: 'What do old people smell like?',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 58,
+    numAnswers: 1,
+    official: true,
+    text: 'The class field trip was completely ruined by _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 59,
+    numAnswers: 1,
+    official: true,
+    text: 'When Pharaoh remained unmoved, Moses called down a Plague of _.',
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 60,
+    numAnswers: 1,
+    official: true,
+    text: "What's my secret power?",
+    region: 'America'
+  },
+  {
+    expansion: 'Base',
+    id: 61,
+    numAnswers: 1,
+    official: true,
+    text: "What's there a ton of in heaven?",
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 62,
+    numAnswers: 1,
+    official: true,
+    text: 'What would grandma find disturbing, yet oddly charming?',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 63,
+    numAnswers: 2,
+    official: true,
+    text: 'I never truly understood _ until I encountered _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 64,
+    numAnswers: 1,
+    official: true,
+    text: 'What did the U.S. airdrop to the children of Afghanistan?',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 65,
+    numAnswers: 1,
+    official: true,
+    text: 'What helps Obama unwind?',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 66,
+    numAnswers: 1,
+    official: true,
+    text: 'What did Vin Diesel eat for dinner?',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 67,
+    numAnswers: 1,
+    official: true,
+    text: '_: good to the last drop.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 68,
+    numAnswers: 1,
+    official: true,
+    text: 'Why am I sticky?',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 69,
+    numAnswers: 1,
+    official: true,
+    text: 'What gets better with age?',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 70,
+    numAnswers: 1,
+    official: true,
+    text: '_: kid-tested, mother-approved.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 71,
+    numAnswers: 1,
+    official: true,
+    text: "What's the crustiest?",
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 72,
+    numAnswers: 1,
+    official: true,
+    text: "What's Teach for America using to inspire inner city students to succeed?",
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 73,
+    numAnswers: 1,
+    official: true,
+    text: 'Studies show that lab rats navigate mazes 50% faster after being exposed to _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 74,
+    numAnswers: 1,
+    official: true,
+    text: 'Life for American Indians was forever changed when the White Man introduced them to _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 76,
+    numAnswers: 1,
+    official: true,
+    text: 'I do not know with what weapons World War III will be fought, but World War IV will be fought with _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 77,
+    numAnswers: 1,
+    official: true,
+    text: 'Why do I hurt all over?',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 78,
+    numAnswers: 1,
+    official: true,
+    text: 'What am I giving up for Lent?',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 79,
+    numAnswers: 1,
+    official: true,
+    text: "In Michael Jackson's final moments, he thought about _.",
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 80,
+    numAnswers: 1,
+    official: true,
+    text: 'In an attempt to reach a wider audience, the Smithsonian Museum of Natural History has opened an interactive exhibit on _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 81,
+    numAnswers: 1,
+    official: true,
+    text: 'When I am President of the United States, I will create the Department of _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 82,
+    numAnswers: 2,
+    official: true,
+    text: 'Lifetime® presents _, the story of _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 83,
+    numAnswers: 1,
+    official: true,
+    text: 'When I am a billionaire, I shall erect a 50-foot statue to commemorate _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 84,
+    numAnswers: 2,
+    official: true,
+    text: 'When I was tripping on acid, _ turned into _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 85,
+    numAnswers: 2,
+    official: true,
+    text: "That's right, I killed _.  How, you ask?  _.",
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 86,
+    numAnswers: 1,
+    official: true,
+    text: "What's my anti-drug?",
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 88,
+    numAnswers: 1,
+    official: true,
+    text: 'What never fails to liven up the party?',
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 89,
+    numAnswers: 1,
+    official: true,
+    text: "What's the new fad diet?",
+    region: 'Europe'
+  },
+  {
+    expansion: 'Base',
+    id: 90,
+    numAnswers: 1,
+    official: true,
+    text: 'Major League Baseball has banned _ for giving players an unfair advantage.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 91,
+    numAnswers: 1,
+    official: true,
+    text: 'My plan for world domination begins with _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 92,
+    numAnswers: 1,
+    official: true,
+    text: 'The CIA now interrogates enemy agents by repeatedly subjecting them to _.',
+    region: 'Europe'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 93,
+    numAnswers: 2,
+    official: true,
+    text: 'Dear Sir or Madam, We regret to inform you that the Office of _ has denied your request for _',
+    region: 'Europe'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 94,
+    numAnswers: 1,
+    official: true,
+    text: 'In Rome, there are whisperings that the Vatican has a secret room devoted to _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 95,
+    numAnswers: 1,
+    official: true,
+    text: 'Science will never explain _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 96,
+    numAnswers: 1,
+    official: true,
+    text: 'When all else fails, I can always masturbate to _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 97,
+    numAnswers: 1,
+    official: true,
+    text: "I learned the hard way that you can't cheer up a grieving friend with _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 98,
+    numAnswers: 1,
+    official: true,
+    text: 'In its new tourism campaign, Detroit proudly proclaims that it has finally eliminated _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 99,
+    numAnswers: 2,
+    official: true,
+    text: 'An international tribunal has found _ guilty of _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 100,
+    numAnswers: 1,
+    official: true,
+    text: 'The socialist governments of Scandinavia have declared that access to _ is a basic human right.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 101,
+    numAnswers: 1,
+    official: true,
+    text: 'In his new self-produced album, Kanye West raps over the sounds of _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 102,
+    numAnswers: 1,
+    official: true,
+    text: "What's the gift that keeps on giving?",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 103,
+    numAnswers: 1,
+    official: true,
+    text: 'Next season on Man vs. Wild, Bear Grylls must survive in the depths of the Amazon with only _ and his wits.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 104,
+    numAnswers: 1,
+    official: true,
+    text: 'When I pooped, what came out of my butt?',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 105,
+    numAnswers: 1,
+    official: true,
+    text: "In the distant future, historians will agree that _ marked the beginning of America's decline.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 106,
+    numAnswers: 2,
+    official: true,
+    text: 'In a pinch, _ can be a suitable substitute for _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 107,
+    numAnswers: 1,
+    official: true,
+    text: 'What has been making life difficult at the nudist colony?',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 108,
+    numAnswers: 2,
+    official: true,
+    text: "Michael Bay's new three-hour action epic pits _ against _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 109,
+    numAnswers: 1,
+    official: true,
+    text: "And I would have gotten away with it, too, if it hadn't been for _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe1',
+    id: 110,
+    numAnswers: 1,
+    official: true,
+    text: 'What brought the orgy to a grinding halt?',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 111,
+    numAnswers: 1,
+    official: true,
+    text: 'During his midlife crisis, my dad got really into _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 112,
+    numAnswers: 2,
+    official: true,
+    text: '_ would be woefully incomplete without _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 113,
+    numAnswers: 1,
+    official: true,
+    text: 'My new favorite porn star is Joey _ McGee.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 114,
+    numAnswers: 1,
+    official: true,
+    text: 'Before I run for president, I must destroy all evidence of my involvement with _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 115,
+    numAnswers: 1,
+    official: true,
+    text: 'This is your captain speaking. Fasten your seatbelts and prepare for _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 116,
+    numAnswers: 1,
+    official: true,
+    text: 'In his newest and most difficult stunt, David Blaine must escape from _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 117,
+    numAnswers: 1,
+    official: true,
+    text: 'The Five Stages of Grief: denial, anger, bargaining, _, and acceptance.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 118,
+    numAnswers: 2,
+    official: true,
+    text: 'My mom freaked out when she looked at my browser history and found _.com/_.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 120,
+    numAnswers: 1,
+    official: true,
+    text: "Members of New York's social elite are paying thousands of dollars just to experience _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 122,
+    numAnswers: 1,
+    official: true,
+    text: 'Little Miss Muffet Sat on a tuffet, Eating her curds and _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 123,
+    numAnswers: 2,
+    official: true,
+    text: "If God didn't want us to enjoy _, he wouldn't have given us _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 124,
+    numAnswers: 1,
+    official: true,
+    text: "My country, 'tis of thee, sweet land of _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 125,
+    numAnswers: 1,
+    official: true,
+    text: 'After months of debate, the Occupy Wall Street General Assembly could only agree on More _!',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 126,
+    numAnswers: 2,
+    official: true,
+    text: 'I spent my whole life working toward _, only to have it ruined by _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 127,
+    numAnswers: 1,
+    official: true,
+    text: 'Next time on Dr. Phil: How to talk to your child about _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 128,
+    numAnswers: 1,
+    official: true,
+    text: 'Only two things in life are certain: death and _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 129,
+    numAnswers: 1,
+    official: true,
+    text: "Everyone down on the ground! We don't want to hurt anyone. We're just here for _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 130,
+    numAnswers: 1,
+    official: true,
+    text: 'The healing process began when I joined a support group for victims of _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 131,
+    numAnswers: 1,
+    official: true,
+    text: 'The votes are in, and the new high school mascot is _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 132,
+    numAnswers: 1,
+    official: true,
+    text: 'Charades was ruined for me forever when my mom had to act out _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 133,
+    numAnswers: 2,
+    official: true,
+    text: 'Before _, all we had was _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe2',
+    id: 135,
+    numAnswers: 2,
+    official: true,
+    text: "You haven't truly lived until you've experienced _ and _ at the same time.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 181,
+    numAnswers: 1,
+    official: true,
+    text: "After blacking out during New year's Eve, I was awoken by _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 182,
+    numAnswers: 1,
+    official: true,
+    text: 'This holiday season, Tim Allen must overcome his fear of _ to save Christmas.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 183,
+    numAnswers: 1,
+    official: true,
+    text: 'Jesus is _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 184,
+    numAnswers: 1,
+    official: true,
+    text: 'Every Christmas, my uncle gets drunk and tells the story about _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 185,
+    numAnswers: 1,
+    official: true,
+    text: 'What keeps me warm during the cold, cold, winter?',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 186,
+    numAnswers: 1,
+    official: true,
+    text: 'On the third day of Christmas, my true love gave to me: three French hens, two turtle doves, and _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHxmas',
+    id: 187,
+    numAnswers: 1,
+    official: true,
+    text: 'Wake up, America. Christmas is under attack by secular liberals and their _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 236,
+    numAnswers: 1,
+    official: true,
+    text: 'A successful job interview begins with a firm handshake and ends with _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 237,
+    numAnswers: 1,
+    official: true,
+    text: "Lovin' you is easy 'cause you're _.",
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 238,
+    numAnswers: 2,
+    official: true,
+    text: 'My life is ruled by a vicious cycle of _ and _.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 240,
+    numAnswers: 1,
+    official: true,
+    text: '_. Awesome in theory, kind of a mess in practice.',
+    region: 'Africa'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 239,
+    numAnswers: 1,
+    official: true,
+    text: 'The blind date was going horribly until we discovered our shared interest in _.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 241,
+    numAnswers: 1,
+    official: true,
+    text: "I'm not like the rest of you. I'm too rich and busy for _.",
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 242,
+    numAnswers: 1,
+    official: true,
+    text: 'In the seventh circle of Hell, sinners must endure _ for all eternity.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 243,
+    numAnswers: 2,
+    official: true,
+    text: '_: Hours of fun. Easy to use. Perfect for _!',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 244,
+    numAnswers: 1,
+    official: true,
+    text: 'What left this stain on my couch?',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 245,
+    numAnswers: 1,
+    official: true,
+    text: 'Call the law offices of Goldstein \u0026 Goldstein, because no one should have to tolerate _ in the workplace.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 246,
+    numAnswers: 2,
+    official: true,
+    text: 'When you get right down to it, _ is just _.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 247,
+    numAnswers: 1,
+    official: true,
+    text: 'Turns out that _-Man was neither the hero we needed nor wanted.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 248,
+    numAnswers: 1,
+    official: true,
+    text: 'As part of his daily regimen, Anderson Cooper sets aside 15 minutes for _.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 249,
+    numAnswers: 1,
+    official: true,
+    text: "Money can't buy me love, but it can buy me _.",
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 250,
+    numAnswers: 2,
+    official: true,
+    text: 'With enough time and pressure, _ will turn into _.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 251,
+    numAnswers: 1,
+    official: true,
+    text: 'And what did you bring for show and tell?',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 252,
+    numAnswers: 1,
+    official: true,
+    text: 'During high school, I never really fit in until I found _ club.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 253,
+    numAnswers: 1,
+    official: true,
+    text: "Hey, baby, come back to my place and I'll show you _.",
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 254,
+    numAnswers: 2,
+    official: true,
+    text: "After months of practice with _, I think I'm finally ready for _.",
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 255,
+    numAnswers: 1,
+    official: true,
+    text: 'To prepare for his upcoming role, Daniel Day-Lewis immersed himself in the world of _.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 256,
+    numAnswers: 1,
+    official: true,
+    text: 'Finally! A service that delivers _ right to your door.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 257,
+    numAnswers: 1,
+    official: true,
+    text: 'My gym teacher got fired for adding _ to the obstacle course.',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 258,
+    numAnswers: 2,
+    official: true,
+    text: 'Having problems with _? Try _!',
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 259,
+    numAnswers: 1,
+    official: true,
+    text: "As part of his contract, Prince won't perform without _ in his dressing room.",
+    region: 'America'
+  },
+  {
+    expansion: 'CAHe3',
+    id: 260,
+    numAnswers: 2,
+    official: true,
+    text: "Listen, son. If you want to get involved with _, I won't stop you. Just steer clear of _.",
+    region: 'America'
+  }
+];
+
+module.exports = questions;
+
